move caf file to caf folder
