--- back/api/test/resources/out.pdf
@@ -1,1 +1,0 @@
-undefined
