+++ back/package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "pretest": "npm install",
-    "test": "istanbul cover --print details _mocha -- --recursive -R spec test",
+    "test": "istanbul cover --print details _mocha -- -R spec api/test/** api/*/test/** --recursive",
     "start": "node ./bin/www"
   },
   "author": "SGMAP",
